refactor: move some session helpers to `@repo/helper`
