+++ packages/helper/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "@repo/helper",
+  "version": "0.0.0",
+  "license": "MIT",
+  "private": true,
+  "exports": {
+    "./session": "./src/session-helper.ts"
+  },
+  "dependencies": {
+    "@repo/config": "workspace:*",
+    "@repo/db": "workspace:*"
+  },
+  "devDependencies": {
+    "@repo/schema": "workspace:*",
+    "@repo/ts-config": "workspace:*",
+    "@types/node": "^22.10.4"
+  }
+}
